Remove progressive disclosure system, simplify home screen
The 5-level progressive disclosure (void → firstLight → gridAwakens →
viewsEmerge → fullPower) forced every UI change to be duplicated across
multiple views. Replaced with: 0 entries = empty state, 1+ = home view.

- Rename HomeAIComposedView → HomeView with inline empty state
- Delete HomeSparseView, VoidView, SettingsMinimalView
- Strip DisclosureLevel enum and all level tracking from AppState
- Simplify RootView routing to single HomeView + SettingsFullView
- Clean up DevMode (remove level picker, keep state toggles)
