+++ Murmur/MurmurApp.swift
@@ -10,19 +10,6 @@ struct MurmurApp: App {
     @State private var notificationPreferences = NotificationPreferences()
     @Environment(\.scenePhase) private var scenePhase
 
-    init() {
-        #if DEBUG
-        // Support launch argument: -disclosureLevel <0-4>
-        if let idx = ProcessInfo.processInfo.arguments.firstIndex(of: "-disclosureLevel"),
-           idx + 1 < ProcessInfo.processInfo.arguments.count,
-           let rawValue = Int(ProcessInfo.processInfo.arguments[idx + 1]),
-           let level = DisclosureLevel(rawValue: rawValue) {
-            appState.devOverrideLevel = level
-            appState.isDevMode = true
-        }
-        #endif
-    }
-
     var body: some Scene {
         WindowGroup {
             RootView()
